requirements.txt
Django<4
djangorestframework
djangorestframework-simplejwt
drf-spectacular
edx-rbac
